debug

+++ nrscope/hdr/dci_decoder.h
@@ -37,7 +37,7 @@ class DCIDecoder{
     bool hidden_bwp = false; 
     uint8_t possible_coreset_total_num;
     // Each possible hidden coreset central freq location
-    std::vector <int> coreset_central_freqs;
+    std::vector <long double> coreset_central_freqs;
     // Ring buf for each possible coreset that track decoded dci num for last 1000 slots
     std::vector <std::vector <uint32_t>> dl_dci_num_1000_tracker;
     std::vector <std::vector <uint32_t>> ul_dci_num_1000_tracker;
